+++ src/app/lhc-form/lhc-form.component.ts
@@ -120,7 +120,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy {
                 q = LForms.FHIR[fhirVer].SDC.convertQuestionnaireToLForms(q);
               }
             }
-            self.lhcFormData = new LhcFormData(q)
+            self.lhcFormData = new LhcFormData(q);
             // and options change
             if (changes.options && self.options) {
               // self.lhcFormData.setTemplateOptions(CommonUtils.deepCopy(self.options));
@@ -130,6 +130,7 @@ export class LhcFormComponent implements OnInit, OnChanges, OnDestroy {
             // if FHIR libs are loaded and the data is converted from a FHIR Questionnaire
             if (LForms.FHIR && self.lhcFormData.fhirVersion) {
               self.lhcFormData.loadFHIRResources(self.prepop).then(()=> {
+                self.lhcFormData._fhirResourceLoaded = true;
                 // when a new form is loaded, run all FHIR Expressions including the initial expressions
                 // self.lhcFormData sometimes is set to null to clear the page
                 if (self.lhcFormData && (self.lhcFormData._hasResponsiveExpr || self.lhcFormData._hasInitialExpr)) {