+++ src/lib/lforms/lhc-form-data.ts
@@ -124,6 +124,7 @@ export default class LhcFormData {
   _showErrors;
   _showWarnings;
   _showInfo;
+  _fhirResourceLoaded;
 
   /**
    * Constructor
@@ -2898,14 +2899,21 @@ export default class LhcFormData {
             else {
               found = false;
             }
-            // a saved value or a default value is not on the list (default answer must be one of the answer items).
-            // non-matching value objects are kept, (data control or others might use data on these objects)
-            if (!found && item.dataType === CONSTANTS.DATA_TYPE.CWE && userValue) {
-              // need a new copy of the data to trigger a change in the data model in autocompleter component
-              // where if the dataModel is in the changes, its value will be preserved when recreating the autocompleter
-              var userValueCopy = CommonUtils.deepCopy(userValue);
-              userValueCopy._notOnList = true;  // _notOnList might have been set above when the orginal value is a string
-              listVals.push(userValueCopy);
+            if (!found && userValue) {
+              // a saved value or a default value is not on the list (default answer must be one of the answer items).
+              // non-matching value objects are kept, (data control or others might use data on these objects)
+              if (item.dataType === CONSTANTS.DATA_TYPE.CWE) {
+                // need a new copy of the data to trigger a change in the data model in autocompleter component
+                // where if the dataModel is in the changes, its value will be preserved when recreating the autocompleter
+                var userValueCopy = CommonUtils.deepCopy(userValue);
+                userValueCopy._notOnList = true;  // _notOnList might have been set above when the orginal value is a string
+                listVals.push(userValueCopy);
+              }
+              // while FHIR Resources are still being loaded, keep the original values
+              // the value will be reset once the the answers are loaded from answerValueSet
+              else if (!this._fhirResourceLoaded) {
+                listVals.push(userValue);
+              }
             }
           }
         }